requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.10.0+cu128
torchvision==0.25.0+cu128
torchaudio==2.10.0+cu128

pytorch-lightning==1.9.1
torchmetrics==1.8.2
tensorboard==2.20.0
tqdm==4.65.2
tabulate==0.8.10
matplotlib==3.10.8
# numpy==2.2.6
# scipy==1.15.2
pyyaml==6.0.3
yacs==0.1.8
packaging==24.2
termcolor==1.1.0
opencv-python==4.10.0.84
pillow==12.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
h5py==3.13.0

smplx==0.1.28
# chumpy==0.70
ftfy==6.3.1
regex==2026.1.15
requests==2.32.5

mmengine==0.10.7
mmcv==1.6.2

git+https://github.com/openai/CLIP.git